Updated Controllers

+++ src/main/java/com/foundgrounds/shops/ShopRepository.java
@@ -2,7 +2,9 @@ package com.foundgrounds.shops;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopRepository extends JpaRepository<Shop, Integer> {
+import com.foundgrounds.shops.Shop;
+
+public interface ShopRepository extends JpaRepository<Shop, Long> {
 
 	
 }
